Add moar week 6 exercises

+++ app/helpers/application_helper.rb
@@ -2,10 +2,13 @@ module ApplicationHelper
   def edit_and_destroy_buttons(item)
     return if current_user.nil?
 
+    del = ''
+    if current_user.admin
+      del = button_to('Destroy', item,  method: :delete,
+                                        form: { data: { turbo_confirm: "Are you sure ?" } },
+                                        class: "btn btn-danger")
+    end
     edit = link_to('Edit', url_for([:edit, item]), class: "btn btn-primary")
-    del = button_to('Destroy', item,  method: :delete,
-                                      form: { data: { turbo_confirm: "Are you sure ?" } },
-                                      class: "btn btn-danger")
     raw("#{edit} #{del}")
   end
